String to Date conversion for ETL from LZ to >IZ
ETL from LandingZone to IntegrationZone now
converts dates from the datatype string to
a correct Date datatype by parsing it accordingly
using Jaspersoft's internal local conversion in
a tMap Action.

Beware: Due to the modifications in the database
scheme of the Integrationzone accomodating the
conversion, the ETL from IntegrationZone to
Core Data Warehouse is broken when based on the new
scheme. You can still use it by dropping and recreating
the database with the SQL dump.

Note: The database server in the ETL Jobs for the
Integration Zone were changed from localhost to
192.168.57.250 because I run the MariaDB Server in a
Virtual Machine and needed to test it.

Furthermore most of the LoadDTS and RecSrc Fields in
the integration zone seem to be missing for yet unknown
reason.

Signed-off-by: Sascha Scherrer <dev@saschascherrer.de>